+++ CHANGELOG.md
@@ -3,6 +3,16 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [v1.3.1] - 2021-01-04
+### Added
+- Collection: long press and move to select/deselect multiple entries
+- Info: show Spherical Video V1 metadata
+- Info: metadata search
+
+### Fixed
+- Viewer: fixed panning inertia following double-tap scaling
+- Collection: fixed crash when loading TIFF files on Android 11
+
 ## [v1.3.0] - 2020-12-26
 ### Added
 - Viewer: quick scale (aka one finger zoom)